Move scroll_behavior into widgets
Previously this code was in the animation layer, which didn't make much
sense because scrolling is a widget concern.
