+++ packages/flutter/lib/animation.dart
@@ -13,6 +13,5 @@ export 'src/animation/animations.dart';
 export 'src/animation/curves.dart';
 export 'src/animation/forces.dart';
 export 'src/animation/listener_helpers.dart';
-export 'src/animation/scroll_behavior.dart';
 export 'src/animation/ticker.dart';
 export 'src/animation/tween.dart';
