+++ packages/flutter/lib/src/widgets/scrollable_grid.dart
@@ -4,12 +4,12 @@
 
 import 'dart:math' as math;
 
+import 'package:flutter/rendering.dart';
+
 import 'framework.dart';
+import 'scroll_behavior.dart';
 import 'scrollable.dart';
 import 'virtual_viewport.dart';
-
-import 'package:flutter/animation.dart';
-import 'package:flutter/rendering.dart';
 
 /// A vertically scrollable grid.
 ///
