+++ packages/flutter/test/widget/remember_scroll_position_test.dart
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
 import 'package:test/test.dart';
